Refactor logging helpers into separate module
